Home screen set up.  All navigation appears to be working as expected.
TODO:  YelpAPI search using coords to get restaurants with blacklisted categories within chosen proximity
TODO:  Settings are updating label, but not persisting -- these need to be stored in UserDefaults.standard and dynamically loaded on viewDidLoad().  We also will need these for generating our notifications.
TODO:  Decide whether we are having users/passwords or not -- if so we'll need to store this info with their UserDefaults.standard variables